clients/python/percolate/pyproject.toml
[tool.poetry]
name = "percolate-db"
version = "0.3.3"
description = "A client for working with the [Percolate database](https://github.com/Percolation-Labs/percolate)"
authors = ["mr-saoirse <amartey@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "percolate"}]

[tool.poetry.dependencies]
apscheduler = "^3.10.1"
httpx = ">=0.28.1"
httpcore = "^1.0.0"
python = ">=3.10,<4.0"
pydantic = ">=2.0.0"
psycopg2-binary = ">=2.0.0"
tenacity = ">=8.0.0"
loguru = ">=0.7.3"
openai = ">=1.0"
docstring-parser = ">=0.16"
typer = ">=0.11.0"
pyyaml = ">=6.0.2"
hypercorn = ">=0.17.3"
requests = ">=2.32.3"
fastapi = ">=0.110.1"
minio = ">=7.2.15"
python-multipart = ">0"
html2text = "^2024.2.26"
# authlib dependency removed as we now use custom implementation
itsdangerous = "^2.2.0"
duckdb = "^0.10.0"
polars = "^0.20.5"
kuzu = "^0.0.11"
pyarrow = "^19.0.1"
aiobotocore = "*"
boto3 = "^1.26.0"  
aiohttp = "^3.8"
mammoth = "^1.9.0"
markdown = "^3.8"
pyjwt = "^2.10.1"
pymupdf = "1.24.13"
pdf2image = "*"
fastmcp = ">=2.10.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pytest-asyncio = "^0.24.0"
pydantic-settings = "^2.0.0"
uvicorn = ">=0.17.0"

[tool.poetry.group.iceberg]
optional = true

[tool.poetry.group.iceberg.dependencies]
pyiceberg = "^0.6.0"


[tool.poetry.group.dev.dependencies]
toml = "^0.10.2"
pytest = "^8.3.5"
pytest-asyncio = "^0.24.0"
pytest-cov = "^4.1.0"
docker = "^6.1.3"

[tool.pytest.ini_options]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
p8 = "percolate.cli:app"

